Changed parameter type from type to recorderType (type is reserved word). Refactored recorders into single file/module.
