+++ .gitignore
@@ -90,3 +90,8 @@ ENV/
 
 # PyCharm
 .idea
+.vscode/*
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
