+++ Recorders.py
@@ -1,4 +1,26 @@
-import Recorder, Measurement
+from Measurement import Measurement
+
+class Recorder(object):
+    def __init__(self, recorderType):
+        self.recorderType = recorderType
+
+    def record(self, measure: Measurement):
+        None
+
+class PrintRecorder(Recorder):
+    def __init__(self, config):
+        Recorder.__init__(self, 'file')
+
+        self.format = config.format
+
+    def record(self, measure: Measurement):
+        line = self.format.format(
+            device_id=measure.device_id,
+            celsius=measure.get_celsius(),
+            fahrenheit=measure.get_fahrenheit(),
+            timestamp=measure.timestamp)
+
+        print(line, end='\n')
 
 class FileRecorder(Recorder):
     def __init__(self, config):
